# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Teens Party Morocco API
  description: |
    API documentation for Teens Party Morocco platform.

    ## Authentication
    Most endpoints require authentication via session cookies (NextAuth).
    Admin endpoints require elevated permissions.

    ## Rate Limiting
    - Standard users: 100 requests/minute
    - Premium users: 500 requests/minute
    - API keys: 1000 requests/minute

    ## Error Handling
    All errors follow the format:
    ```json
    {
      "success": false,
      "error": "Error message"
    }
    ```
  version: 1.0.0
  contact:
    name: Teens Party Morocco Support
    email: support@teensparty.ma
    url: https://teensparty.ma/contact
  license:
    name: Proprietary
    url: https://teensparty.ma/terms
servers:
  - url: https://teensparty.ma/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /payments/xp:
    post:
      tags:
        - Payments
      summary: Pay with XP coins
      description: Use accumulated XP coins for payment
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bookingId
                - amount
              properties:
                bookingId:
                  type: string
                  format: uuid
                amount:
                  type: number
                  minimum: 0
      responses:
        '200':
          description: XP payment processed
...
  /parent/teens/create:
    post:
      tags:
        - Parent
      summary: Create teen profile
      description: Creates a new teen profile linked to parent
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeenRequest'
      responses:
        '200':
          description: Teen profile created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Teen'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
  schemas:
...
    Teen:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pseudo:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        avatarUrl:
          type: string
          format: uri
    CreateTeenRequest:
      type: object
      required:
        - pseudo
        - firstName
        - lastName
        - dateOfBirth
      properties:
        pseudo:
          type: string
          minLength: 3
          maxLength: 20
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        school:
          type: string
        interests:
          type: array
          items:
            type: string
...